Move pkg/k8scontext/event.go to pkg/events/events.go
